+++ TTS/tts/layers/delightful_tts/phoneme_prosody_predictor.py
@@ -3,7 +3,7 @@ from typing import Dict, Any
 import torch
 import torch.nn as nn
 
-from TTS.tts.layers.d_tts.conv_layers import ConvTransposed
+from TTS.tts.layers.delightful_tts.conv_layers import ConvTransposed
 
 class PhonemeProsodyPredictor(nn.Module):
     """Non-parallel Prosody Predictor inspired by: https://arxiv.org/pdf/2102.00851.pdf
